requirements-dev.txt
# Инструменты для разработки
# Установите их для проверки качества кода:
# pip install -r requirements-dev.txt

# Линтеры и форматтеры
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Типы для Django
django-stubs>=4.2.0
types-python-dateutil>=2.8.19

# Тестирование (для будущего)
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
